requirements-dev.txt
# For developers wanting to contribute to AudioSeal
func_argparse
torchaudio
soundfile
pytest
mypy
black
isort
flake8
pre-commit
huggingface_hub
wavmark